server/requirements.txt
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Testing==0.6.2
gunicorn>=19.7.1,<20.0.0
gevent>=1.2.2<2.0.0
adafruit-pca9685==1.0.1
